requirements.txt
kivy  # pip3 install kivy
opencv-python  # pip3 install opencv-python   - cv2 is part of it
# sudo apt-get install zbar-tools
pyzbar  # pip3 install pyzbar
python-dotenv  # pip3 install python-dotenv
sqlalchemy  # pip3 install sqlalchemy
pyaml  # pip3 install pyaml
pyqrcode  # pip3 install pyqrcode
bitcoinlib      # needed - bitcoin management
requests        # needed for Node-calls
